// src/app/services/model/performanceIndicatorModel.ts
/**
 * Metal Casting Project API
 * This API serves the frontend with data from databases and other external services.
 *
 * The version of the OpenAPI document: v1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface PerformanceIndicatorModel { 
    title?: string | null;
    overallTarget?: number;
    category?: string | null;
    actualOverallTarget?: number;
    firstQuarterTarget?: number;
    firstQuarterActual?: number;
    secondQuarterTarget?: number;
    secondQuarterActual?: number;
    thirdQuarterTarget?: number;
    thirdQuarterActual?: number;
    fourthQuarterTarget?: number;
    fourthQuarterActual?: number;
}